solution/python/requirements.txt
flask==2.3.2
requests==2.31.0

# For version compatibility issues
opentelemetry-distro==0.40b0
opentelemetry-sdk==1.19.0
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0